Added KA-50 BlackShark (see note)
This is an operational build but not yet completed. The
'wac_hc_base_twin' is still in progress and has strange errors with any
twin bladed aircraft that is not coaxial. Do not try and make a twin
bladed chopper with this until it is completed! Apart from that the
blackshark itself works fine.